src/tests/requirements-test.txt
# Test dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development tools
black==23.3.0
flake8==6.0.0
mypy==1.3.0

# SC2 API
python-sc2==1.0.0

# For type checking
typing-extensions==4.5.0
